# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Garden Living API
  version: 0.1.0
servers:
  - url: https://api.gardenliving.app
paths:
  /v1/tasks/today:
    get:
      summary: List today's tasks
      parameters:
        - in: query
          name: timezone
          schema:
            type: string
      responses:
        '200':
          description: Task list
  /v1/tasks/{taskId}/complete:
    post:
      summary: Complete a task
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                offlineCreatedAt:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Completion recorded
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []